Added Space Checking
Space Checking refers to Spaces that are must after the keywords. Since a non space after the keywords means that its not a keyword. It might be an identifier or an invalid type.  I also put "None equation" first on the last if loop condition of every keywords since program might run the "in statement" using a "None" operand FIRST, which would result in TypeError.